remove unnecessary cast

+++ utils/Arduino/gpioHandler.cpp
@@ -6,20 +6,20 @@
 
 bool GPIO::getPort(uint8_t port)
 {
-    return digitalRead((uint8_t)port);
+    return digitalRead(port);
 }
 
 uint16_t GPIO::getAnalogPort(uint8_t port)
 {
-    return analogRead((uint8_t)port);
+    return analogRead(port);
 }
 
 void GPIO::setPort(uint8_t port, bool value)
 {
-    digitalWrite((uint8_t)port, value);
+    digitalWrite(port, value);
 }
 
 void GPIO::setAnalogPort(uint8_t port, uint8_t value)
 {
-    analogWrite((uint8_t)port, value);
+    analogWrite(port, value);
 }
